requirements.txt
cffi==1.15.1
click==8.1.3
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
engi1020==5.1.2
fonttools==4.39.2
hexdump==3.3
kiwisolver==1.4.4
maskpass==0.3.7
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pycparser==2.21
pynput==1.7.6
pyobjc-core==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Quartz==9.0.1
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
regex==2023.3.23
six==1.16.0